requirements.txt
autobahn[twisted]
Flask
pyserial
twisted-mqtt
pyudev

# Autobahn dependencies
six
txaio
zope.interface
twisted
setuptools
Automat
constantly
incremental
appdirs
packaging
attrs
pyparsing

# Flask dependencies
itsdangerous
click
Werkzeug
MarkupSafe
Jinja2
